+++ Rayson/src/org/rayson/transport/client/RpcConnection.java
@@ -195,7 +195,7 @@ class RpcConnection extends PacketConnection {
 	}
 
 	@Override
-	public int getVersion() {
+	public short getVersion() {
 
 		return version;
 	}
